Split RunRecorder metrics helper tests
