+++ apps/server/tests/use_cases/run/test_metrics_log_helpers.py
@@ -1,32 +1,13 @@
 from __future__ import annotations
 
-import asyncio
-import logging
-import threading
-import time
-from pathlib import Path
-from types import SimpleNamespace
-
 import pytest
-from test_support.core import wait_until
-from test_support.persisted_analysis import make_persisted_analysis
 
-from vibesensor.adapters.persistence.history_db import create_history_persistence_adapters
-from vibesensor.domain import AnalysisSettingsSnapshot, CarSnapshot
-from vibesensor.shared.types.history_records import AnalyzingRunHealth
-from vibesensor.use_cases.run import _recorder_runtime
-from vibesensor.use_cases.run._recorder_types import _build_run_metadata_record
 from vibesensor.use_cases.run.lifecycle_state import ActiveRunSnapshot
-from vibesensor.use_cases.run.post_analysis import PostAnalysisHealthSnapshot
 
 
-class _NullDB:
-    """Stub DB for tests that need a non-None history_db without real DB ops."""
-
-    def analyzing_run_health(self):
-        return AnalyzingRunHealth(analyzing_run_count=0, analyzing_oldest_age_s=None)
-
-
+# Remaining private seam: these tests need a deterministic active recorder plus
+# exactly one flush without running the async loop. Assertions stay on emitted
+# samples, status, and persisted output rather than recorder object shape.
 def _started_snapshot(logger) -> ActiveRunSnapshot:
     logger.start_recording()
     snapshot = logger._session_snapshot()
@@ -399,608 +380,3 @@ def test_append_records_does_not_timeout_on_brief_gap(
     )
 
     assert timed_out is False
-
-
-def test_stop_recording_does_not_block_on_post_analysis(
-    make_logger,
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Stopping capture should be fast even when analysis is slow.
-
-    Post-analysis belongs in background processing because users expect stop controls to respond
-    promptly. This test simulates a slow summarizer and requires stop_recording to return quickly
-    while analysis completes asynchronously afterward.
-    """
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-    summary_started = threading.Event()
-    allow_summary_finish = threading.Event()
-
-    def _slow_analysis_runner(_run):
-        summary_started.set()
-        assert allow_summary_finish.wait(timeout=5.0)
-        return make_persisted_analysis(
-            {
-                "findings": [],
-                "top_causes": [],
-                "analysis_metadata": {},
-                "case_id": "mock-case",
-            }
-        )
-
-    monkeypatch.setattr(
-        "vibesensor.use_cases.run.logger.build_post_analysis_summary",
-        _slow_analysis_runner,
-    )
-    logger = make_logger(history_db=history_db.run_repository)
-
-    snapshot = _started_snapshot_with_sample(logger)
-    run_id = snapshot.run_id
-
-    started = time.monotonic()
-    logger.stop_recording()
-    elapsed = time.monotonic() - started
-
-    # stop_recording() must return quickly; summary runs in a worker thread.
-    # 5.0s threshold guards against blocking; not a performance target.
-    assert elapsed < 5.0, f"stop_recording() blocked for {elapsed:.2f}s (expected < 5.0s)"
-    assert summary_started.wait(timeout=2.0)
-    allow_summary_finish.set()
-
-    def _status():
-        run = history_db.run_repository.get_run(run_id)
-        return run.status.value if run is not None else None
-
-    assert wait_until(lambda: _status() == "complete", timeout_s=5.0)
-
-
-def test_post_analysis_unexpected_failure_surfaces_worker_error_status(
-    make_logger,
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-
-    def _failing_analysis_runner(_run) -> dict[str, object]:
-        raise RuntimeError("analysis exploded")
-
-    monkeypatch.setattr(
-        "vibesensor.use_cases.run.logger.build_post_analysis_summary",
-        _failing_analysis_runner,
-    )
-    logger = make_logger(history_db=history_db.run_repository)
-
-    snapshot = _started_snapshot_with_sample(logger)
-    run_id = snapshot.run_id
-
-    logger.stop_recording()
-
-    def _status():
-        return logger.status().last_completed_run_error
-
-    expected_worker_bug = "Unexpected post-analysis worker bug: analysis exploded"
-    assert wait_until(lambda: _status() == expected_worker_bug, timeout_s=2.0)
-    status = logger.status()
-    assert status.last_completed_run_error == expected_worker_bug
-    assert status.write_error == f"post-analysis worker bug for run {run_id}: analysis exploded"
-    run = history_db.run_repository.get_run(run_id)
-    assert run is not None
-    assert run.analysis is None
-    assert run.status.value == "error"
-    assert run.error_message == expected_worker_bug
-
-
-def test_post_analysis_burst_uses_single_daemon_worker(
-    make_logger,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    logger = make_logger(history_db=_NullDB())
-
-    active = 0
-    max_active = 0
-    seen: list[str] = []
-    state_lock = threading.Lock()
-
-    def _slow_post_analysis(run_id: str) -> None:
-        nonlocal active, max_active
-        with state_lock:
-            active += 1
-            max_active = max(max_active, active)
-        time.sleep(0.005)
-        seen.append(run_id)
-        with state_lock:
-            active -= 1
-
-    monkeypatch.setattr(logger._post_analysis, "_run_post_analysis", _slow_post_analysis)
-
-    for idx in range(6):
-        logger.schedule_post_analysis(f"run-{idx}")
-
-    logger.wait_for_post_analysis(timeout_s=3.0)
-
-    assert max_active == 1
-    assert len(seen) == 6
-    # After all work completes, the worker thread should have exited and
-    # been cleared to allow a fresh thread on the next scheduling call.
-    with logger._post_analysis._lock:
-        worker = logger._post_analysis._analysis_thread
-    assert worker is None
-
-
-def test_shutdown_blocks_new_start_recording_until_wait_completes(
-    make_logger,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    logger = make_logger(history_db=_NullDB())
-    logger.start_recording()
-    initial_run_id = logger._run_id
-    assert initial_run_id is not None
-
-    allow_wait = threading.Event()
-
-    def _wait(timeout_s: float = 30.0) -> bool:
-        assert timeout_s == 30.0
-        start_result = logger.start_recording()
-        assert start_result.enabled is False
-        assert start_result.run_id is None
-        # Session was stopped by shutdown
-        assert logger._run_id is None
-        allow_wait.set()
-        return True
-
-    monkeypatch.setattr(logger._post_analysis, "wait", _wait)
-
-    assert logger.shutdown() is True
-    assert allow_wait.is_set()
-    restarted = logger.start_recording()
-    assert restarted.enabled is True
-    assert restarted.run_id is not None
-    assert restarted.run_id != initial_run_id
-
-
-def test_shutdown_report_exposes_timeout_state(
-    make_logger,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    logger = make_logger()
-    logger.start_recording()
-
-    monkeypatch.setattr(logger._post_analysis, "wait", lambda timeout_s=30.0: False)
-    monkeypatch.setattr(
-        logger._post_analysis,
-        "snapshot",
-        lambda: PostAnalysisHealthSnapshot(
-            queue_depth=2,
-            active_run_id="run-slow",
-            active_started_at=None,
-            oldest_queued_at=time.time() - 5.0,
-            max_queue_depth=2,
-            last_completed_run_id=None,
-            last_completed_error=None,
-        ),
-    )
-
-    report = logger.shutdown_report(timeout_s=0.1)
-
-    assert report.completed is False
-    assert report.active_run_id_before_stop is not None
-    assert report.analysis_queue_depth == 2
-    assert report.analysis_active_run_id == "run-slow"
-    assert report.final_status.enabled is False
-
-
-def test_stop_recording_continues_when_raw_capture_finalize_degrades(
-    make_logger,
-    fake_history_db,
-    monkeypatch: pytest.MonkeyPatch,
-    caplog: pytest.LogCaptureFixture,
-) -> None:
-    from vibesensor.use_cases.run.raw_capture_writer import RawCaptureFinalizeResult
-
-    scheduled: list[str] = []
-    logger = make_logger(history_db=fake_history_db)
-    snapshot = _started_snapshot_with_sample(logger)
-    logger._raw_capture = SimpleNamespace(
-        finalize_run=lambda run_id, *, sensor_losses=None: RawCaptureFinalizeResult(
-            status="timeout",
-            error="raw capture finalize timed out",
-            queue_depth=3,
-        ),
-        shutdown=lambda timeout_s=5.0: True,
-    )
-    monkeypatch.setattr(logger, "schedule_post_analysis", scheduled.append)
-
-    with caplog.at_level(logging.WARNING, logger="vibesensor.use_cases.run.logger"):
-        status = logger.stop_recording()
-
-    assert status.enabled is False
-    assert fake_history_db.finalize_calls == [snapshot.run_id]
-    updated_run_id, metadata = fake_history_db.updated_metadata[-1]
-    assert updated_run_id == snapshot.run_id
-    assert metadata.raw_capture_finalize is not None
-    assert metadata.raw_capture_finalize.status == "timeout"
-    assert metadata.raw_capture_finalize.queue_depth == 3
-    assert metadata.raw_capture_finalize.error_summary == "raw capture finalize timed out"
-    assert scheduled == []
-    assert "raw_capture_finalize_degraded" in caplog.text
-
-
-class _DelegatingHistoryDB:
-    def __init__(self, repository) -> None:
-        self._repository = repository
-
-    def __getattr__(self, name: str):
-        return getattr(self._repository, name)
-
-
-class _FinalizeSkippedHistoryDB(_DelegatingHistoryDB):
-    async def afinalize_run(
-        self,
-        run_id: str,
-        end_time_utc: str,
-        metadata=None,
-        case_id=None,
-    ) -> bool:
-        return False
-
-
-class _ZeroAppendHistoryDB(_DelegatingHistoryDB):
-    async def aappend_samples(self, run_id: str, samples) -> int:
-        return 0
-
-
-@pytest.mark.parametrize(
-    ("history_db_factory", "raw_capture_status", "expected_statuses", "expected_resolve_reason"),
-    [
-        (
-            lambda repository: repository,
-            "completed",
-            {
-                "FlushPendingRowsStage": "skipped",
-                "FinalizeRawCaptureStage": "ok",
-                "FinalizePersistenceStage": "ok",
-                "ResolvePostAnalysisCandidateStage": "ok",
-            },
-            "ready",
-        ),
-        (
-            lambda repository: repository,
-            "timeout",
-            {
-                "FlushPendingRowsStage": "skipped",
-                "FinalizeRawCaptureStage": "degraded",
-                "FinalizePersistenceStage": "ok",
-                "ResolvePostAnalysisCandidateStage": "skipped",
-            },
-            "raw_capture_finalize_unsettled",
-        ),
-        (
-            _FinalizeSkippedHistoryDB,
-            "completed",
-            {
-                "FlushPendingRowsStage": "skipped",
-                "FinalizeRawCaptureStage": "ok",
-                "FinalizePersistenceStage": "degraded",
-                "ResolvePostAnalysisCandidateStage": "skipped",
-            },
-            "persistence_finalize_unsettled",
-        ),
-        (
-            _ZeroAppendHistoryDB,
-            "not_configured",
-            {
-                "FlushPendingRowsStage": "skipped",
-                "FinalizeRawCaptureStage": "skipped",
-                "FinalizePersistenceStage": "ok",
-                "ResolvePostAnalysisCandidateStage": "skipped",
-            },
-            "history_not_ready",
-        ),
-    ],
-)
-def test_stop_recording_persists_finalization_stages_in_history_metadata(
-    make_logger,
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-    history_db_factory,
-    raw_capture_status: str,
-    expected_statuses: dict[str, str],
-    expected_resolve_reason: str,
-) -> None:
-    from vibesensor.use_cases.run.raw_capture_writer import RawCaptureFinalizeResult
-
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-    logger = make_logger(history_db=history_db_factory(history_db.run_repository))
-    snapshot = _started_snapshot_with_sample(logger)
-    logger._raw_capture = SimpleNamespace(
-        finalize_run=lambda run_id, *, sensor_losses=None: RawCaptureFinalizeResult(
-            status=raw_capture_status,
-            error=("raw capture finalize timed out" if raw_capture_status == "timeout" else None),
-            queue_depth=(3 if raw_capture_status == "timeout" else None),
-        ),
-        shutdown=lambda timeout_s=5.0: True,
-    )
-    monkeypatch.setattr(logger, "schedule_post_analysis", lambda run_id: None)
-
-    logger.stop_recording()
-
-    stored_metadata = history_db.run_repository.get_run_metadata(snapshot.run_id)
-    assert stored_metadata is not None
-    stage_by_name = {stage.stage_name: stage for stage in stored_metadata.finalization_stages}
-    assert {name: stage.status for name, stage in stage_by_name.items()} == expected_statuses
-    assert (
-        stage_by_name["ResolvePostAnalysisCandidateStage"].diagnostic_context["reason"]
-        == expected_resolve_reason
-    )
-
-
-def test_late_raw_capture_finalize_schedules_post_analysis_after_metadata_update(
-    make_logger,
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from vibesensor.use_cases.run.raw_capture_writer import RawCaptureFinalizeResult
-
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-    scheduled: list[str] = []
-    logger = make_logger(history_db=history_db.run_repository)
-    snapshot = _started_snapshot_with_sample(logger)
-    logger._raw_capture = SimpleNamespace(
-        finalize_run=lambda run_id, *, sensor_losses=None: RawCaptureFinalizeResult(
-            status="timeout",
-            error="raw capture finalize timed out",
-        ),
-        shutdown=lambda timeout_s=5.0: True,
-    )
-    monkeypatch.setattr(logger, "schedule_post_analysis", scheduled.append)
-
-    logger.stop_recording()
-    logger._handle_late_raw_capture_finalize_result(
-        snapshot.run_id,
-        RawCaptureFinalizeResult(status="completed"),
-    )
-
-    assert scheduled == [snapshot.run_id]
-    stored_metadata = history_db.run_repository.get_run_metadata(snapshot.run_id)
-    assert stored_metadata is not None
-    assert stored_metadata.raw_capture_finalize is not None
-    assert stored_metadata.raw_capture_finalize.status == "completed"
-
-
-def test_permanent_raw_capture_finalize_failure_schedules_with_degraded_metadata(
-    make_logger,
-    fake_history_db,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from vibesensor.use_cases.run.raw_capture_writer import RawCaptureFinalizeResult
-
-    scheduled: list[str] = []
-    logger = make_logger(history_db=fake_history_db)
-    snapshot = _started_snapshot_with_sample(logger)
-    logger._raw_capture = SimpleNamespace(
-        finalize_run=lambda run_id, *, sensor_losses=None: RawCaptureFinalizeResult(
-            status="failed",
-            error="raw capture finalize failed",
-            queue_depth=0,
-        ),
-        shutdown=lambda timeout_s=5.0: True,
-    )
-    monkeypatch.setattr(logger, "schedule_post_analysis", scheduled.append)
-
-    logger.stop_recording()
-
-    updated_run_id, metadata = fake_history_db.updated_metadata[-1]
-    assert updated_run_id == snapshot.run_id
-    assert metadata.raw_capture_finalize is not None
-    assert metadata.raw_capture_finalize.status == "failed"
-    assert metadata.raw_capture_finalize.error_summary == "raw capture finalize failed"
-    assert scheduled == [snapshot.run_id]
-
-
-def test_post_analysis_uses_run_language_from_metadata(
-    make_logger,
-    tmp_path: Path,
-) -> None:
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-    logger = make_logger(
-        history_db=history_db.run_repository,
-        language_reader=SimpleNamespace(language="nl"),
-    )
-
-    snapshot = _started_snapshot_with_sample(logger)
-    run_id = snapshot.run_id
-
-    def _analysis_runner(run):
-        assert run.run_id == snapshot.run_id
-        assert run.context.language == "nl"
-        assert run.language == "nl"
-        assert run.total_summary_row_count == len(run.samples)
-        assert run.stride == 1
-        return make_persisted_analysis(
-            {
-                "lang": run.language,
-                "row_count": len(run.samples),
-                "analysis_metadata": {
-                    "analyzed_sample_count": len(run.samples),
-                    "total_sample_count": run.total_summary_row_count,
-                    "sampling_method": "full",
-                },
-                "run_suitability": [],
-            }
-        )
-
-    logger._post_analysis._analysis_runner = _analysis_runner
-    logger.stop_recording()
-
-    def _status():
-        run = history_db.run_repository.get_run(run_id)
-        return run.status.value if run is not None else None
-
-    assert wait_until(lambda: _status() == "complete", timeout_s=2.0)
-    stored = history_db.run_repository.get_run(run_id).analysis
-    assert stored is not None
-    assert stored["lang"] == "nl"
-
-
-def test_run_metadata_captures_active_car_snapshot(make_logger) -> None:
-    settings_reader = type(
-        "SettingsReaderStub",
-        (),
-        {
-            "active_car_snapshot": lambda self: CarSnapshot(
-                car_id="car-1",
-                name="Primary",
-                car_type="sedan",
-                aspects={
-                    "tire_width_mm": 255.0,
-                    "tire_aspect_pct": 40.0,
-                    "rim_in": 19.0,
-                    "final_drive_ratio": 3.15,
-                    "current_gear_ratio": 0.81,
-                },
-            ),
-            "analysis_settings_snapshot": lambda self: AnalysisSettingsSnapshot(
-                tire_width_mm=255.0,
-                tire_aspect_pct=40.0,
-                rim_in=19.0,
-                final_drive_ratio=3.15,
-                current_gear_ratio=0.81,
-            ),
-        },
-    )()
-    logger = make_logger(settings_reader=settings_reader)
-
-    metadata = _build_run_metadata_record(logger, "run-1", "2026-01-01T00:00:00Z")
-
-    assert metadata.incomplete_for_order_analysis is False
-    assert metadata.car is not None
-    assert metadata.car.car_id == "car-1"
-    assert metadata.car.name == "Primary"
-    assert metadata.car.car_type == "sedan"
-
-
-def test_run_metadata_captures_recorded_utc_offset(
-    make_logger,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    logger = make_logger()
-    monkeypatch.setattr(
-        "vibesensor.use_cases.run._recorder_types.current_utc_offset_seconds",
-        lambda: 7200,
-    )
-
-    metadata = _build_run_metadata_record(logger, "run-1", "2026-01-01T00:00:00Z")
-
-    assert metadata.recorded_utc_offset_seconds == 7200
-
-
-def test_db_persists_when_jsonl_disabled(make_logger, tmp_path: Path) -> None:
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-    logger = make_logger(history_db=history_db.run_repository, persist_history_db=True)
-
-    snapshot = _started_snapshot_with_sample(logger)
-    run_id = snapshot.run_id
-    logger.stop_recording()
-
-    assert history_db.run_repository.get_run(run_id) is not None
-    assert history_db.run_repository.get_run_samples(run_id)
-
-
-def test_post_analysis_caps_sample_count_and_stores_sampling_metadata(
-    make_logger,
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    # Reduce the cap so we only need ~250 iterations instead of 13 000 (28 s -> <1 s).
-    cap = 200
-    monkeypatch.setattr(
-        "vibesensor.use_cases.run.post_analysis_loader._MAX_POST_ANALYSIS_SAMPLES",
-        cap,
-    )
-
-    history_db = create_history_persistence_adapters(tmp_path / "history.db")
-    logger = make_logger(history_db=history_db.run_repository)
-
-    snapshot = _started_snapshot(logger)
-    run_id = snapshot.run_id
-    start_time_utc = snapshot.start_time_utc
-    start_mono = snapshot.start_mono_s
-    for _ in range(cap + 50):
-        logger._sample_flush.append_records(run_id, start_time_utc, start_mono)
-
-    def _analysis_runner(run):
-        assert run.run_id == snapshot.run_id
-        assert run.language == "en"
-        return make_persisted_analysis(
-            {
-                "row_count": len(run.samples),
-                "analysis_metadata": {
-                    "analyzed_sample_count": len(run.samples),
-                    "total_sample_count": run.total_summary_row_count,
-                    "sampling_method": ("full" if run.stride == 1 else f"stride_{run.stride}"),
-                },
-                "run_suitability": (
-                    [
-                        {
-                            "check_key": "SUITABILITY_CHECK_ANALYSIS_SAMPLING",
-                            "state": "warn",
-                            "explanation": f"stride={run.stride}",
-                        }
-                    ]
-                    if run.stride > 1
-                    else []
-                ),
-            }
-        )
-
-    logger._post_analysis._analysis_runner = _analysis_runner
-    logger.stop_recording()
-
-    def _status():
-        run = history_db.run_repository.get_run(run_id)
-        return run.status.value if run is not None else None
-
-    assert wait_until(lambda: _status() == "complete", timeout_s=3.0)
-    stored = history_db.run_repository.get_run(run_id).analysis
-    assert stored is not None
-    assert stored["row_count"] <= cap
-    assert stored["analysis_metadata"]["total_sample_count"] >= stored["row_count"]
-    assert stored["analysis_metadata"]["sampling_method"].startswith("stride_")
-    suitability_checks = {str(item.get("check_key")) for item in stored.get("run_suitability", [])}
-    assert "SUITABILITY_CHECK_ANALYSIS_SAMPLING" in suitability_checks
-
-
-@pytest.mark.asyncio
-async def test_run_offloads_flush_cycle_with_to_thread(
-    make_logger,
-    fake_history_db,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    logger = make_logger(history_db=fake_history_db)
-
-    logger.start_recording()
-    captured: dict[str, object] = {}
-
-    async def _fake_to_thread(func: object, *args: object, **kwargs: object) -> tuple[None, bool]:
-        captured["func"] = func
-        captured["args"] = args
-        captured["kwargs"] = kwargs
-        return None, False
-
-    async def _cancel_sleep(_interval: float) -> None:
-        raise asyncio.CancelledError()
-
-    monkeypatch.setattr(
-        "vibesensor.use_cases.run._recorder_runtime.asyncio.to_thread",
-        _fake_to_thread,
-    )
-    monkeypatch.setattr(
-        "vibesensor.use_cases.run._recorder_runtime.asyncio.sleep",
-        _cancel_sleep,
-    )
-
-    with pytest.raises(asyncio.CancelledError):
-        await logger.run()
-
-    assert captured.get("func") == _recorder_runtime._flush_active_run_tick
-    assert captured.get("args") == (logger,)
