+++ tools/dev/oversized_test_allowlist.yml
@@ -6,8 +6,6 @@ report_limit: 10
 allowlist:
   # Settings browser regression keeps the moved multi-tab matrix outside smoke.
   apps/ui/tests/regression.settings.spec.ts: Multi-tab settings browser regression matrix moved out of smoke but intentionally stays aggregated.
-  # Metrics log helper coverage still lives as one dense helper-focused module.
-  apps/server/tests/use_cases/run/test_metrics_log_helpers.py: Helper-focused metrics log regression coverage intentionally stays aggregated in one module.
   # Maintenance mounted-DOM coverage now lives in one Node-side regression owner instead of four legacy bridge specs.
   apps/ui/tests/maintenance_feature_signal.spec.ts: Mounted maintenance regression matrix intentionally stays together while replacing the deleted legacy createPanel Playwright split.
   # Cars browser regression keeps the moved car-selection matrix outside smoke.
